TLGPE-07:ADD: Carousel style, UseFetch
