+++ backend/index.js
@@ -22,3 +22,5 @@ app.get("/products", async (req, res) => {
 app.listen(3000, () => {
   console.log("Server in ascolto sulla porta 3000");
 });
+
+// Per avviare il server scrivere "nodemon index.js"
