+++ package.json
@@ -10,11 +10,16 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "body-parser": "^1.20.2",
     "contentful": "^10.9.2",
+    "cors": "^2.8.5",
+    "express": "^4.19.2",
+    "node-fetch": "^3.3.2",
+    "nodemon": "^3.1.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.23.0",
-    "sass": "^1.76.0"
+    "sass": "^1.77.0"
   },
   "devDependencies": {
     "@types/node": "^20.12.9",
