+++ src/components/Carousel/Carousel.scss
@@ -1,0 +1,30 @@
+@import "../../style.scss";
+
+.carousel {
+    display: flex;
+    align-items: center;
+    width: 70rem;
+    .btn-arrow{
+        font-size: large;
+        width: 3rem;
+        height: 3rem;
+        border-radius: 20px;
+        border: 0px;
+    }
+    .carousel-content {
+        display: flex;
+        flex-wrap: nowrap;
+        .carousel-item {
+            width: 100%;
+            max-height: 100%;
+            display: flex;
+            justify-content: space-around;
+            align-items: center;
+            padding: 10px;
+            .carousel-img {
+                width: 100%;
+                height: 100%;
+            }
+        }
+    }
+}
