+++ src/interfaces.ts
@@ -8,3 +8,14 @@ export interface Card {
     img: string;
 }
 
+export interface Products {
+    id: number;
+    name: string;
+    price: number;
+    gender: string;
+    category: string;
+    img: string;
+    quantity: number;
+    description: string;
+  }
+
