+++ src/style.scss
@@ -1,18 +1,16 @@
-*{
+* {
     margin: 0;
     padding: 0;
     box-sizing: border-box;
 }
 
-:root {
-    /* Fonts */
-    --font-family: "Playfair Display", serif;
-    
-    /* Colori */
-    --main-white: #FDF4F4;
-    --secondary-white: #F5EBE0;
-    --main-black: #0E0E0E; 
-}
+/* Fonts */
+$main-font: "Playfair Display", serif;
+/* Colori */
+$main-white: #FDF4F4;
+$econdary-white: #F5EBE0;
+$main-black: #0E0E0E;
+
 
 @font-face {
     font-family: 'Playfair Display';
@@ -24,7 +22,7 @@
 
 /* ProductCard */
 
-.single-card-container{
+.single-card-container {
     display: flex;
     flex-direction: column;
     justify-content: center;
@@ -35,26 +33,26 @@
     gap: 10px;
 }
 
-.card-title{
+.card-title {
     font-weight: bold;
     letter-spacing: 0.1rem;
     width: 300px;
-    
+
 }
 
-.card-body-container{
+.card-body-container {
     // background-color: grey;
     // padding:10px;
     width: 300px;
 }
 
-.card-body-img img{
+.card-body-img img {
     width: 300px;
     height: 450px;
-    object-fit:cover;
+    object-fit: cover;
 }
 
-.card-foot-btn button{
+.card-foot-btn button {
     padding: 5px;
     width: 300px;
     color: white;
@@ -64,10 +62,10 @@
     letter-spacing: 0.5rem;
 }
 
-.card-foot-btn button:hover{
+.card-foot-btn button:hover {
     opacity: 0.3;
 }
 
-.card-body-text{
+.card-body-text {
     width: 300px;
 }
